Fix #108
Parameter tokens were being spaced incompletely.
Now use same token spacing strategy across both parameter parsing and
non-parameter snakefile parsing.
